+++ source/stats/plot-theme.R
@@ -1,8 +1,10 @@
-# Custom themeing function for EDA plots
-eda_plot_theme <- function(p) {
+library(ggplot2)
+
+# Adds sensible axis scales and legend styling to a plot
+add_theme <- function(p) {
 	p +
-	scale_x_continuous(breaks = scales::pretty_breaks(n=10)) +
-	scale_y_continuous(breaks = scales::pretty_breaks(n=10)) +
+	scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
+	scale_y_continuous(breaks = scales::pretty_breaks(n = 10)) +
 	theme_minimal() +
 	theme(
 		plot.title = element_text(size = 20, hjust = 0.5),
@@ -16,4 +18,3 @@ eda_plot_theme <- function(p) {
 		legend.direction = "vertical"
 	)
 }
-